Implemented fixes for read_archive and related tests
