+++ tests/functions/test_complete.py
@@ -432,8 +432,7 @@ def test_complete_multiple_groupings():
         fill_value={"tag_count": 0},
         sort=True,
     ).astype({"tag_count": int})
-    print(result)
-    assert_frame_equal(result, output3, check_dtype=False)
+    assert_frame_equal(result, output3)
 
 
 def test_fill_value_scalar(taxonomy_df):
@@ -452,7 +451,7 @@ def test_fill_value_scalar(taxonomy_df):
         .sort_values("Taxon", ignore_index=True)
     )
 
-    assert_frame_equal(result, expected, check_dtype=False)
+    assert_frame_equal(result, expected)
 
 
 #  http://imachordata.com/2016/02/05/you-complete-me/
