+++ tests/test_documentation_build.py
@@ -25,7 +25,7 @@ def test_docs_general_functions_present():
     # I put in a subsample of general functions.
     # This can be made much more robust.
     rendered_correctly = False
-    with open("./site/api/functions/index.html", "r+", encoding="utf-8") as f:
+    with open("./site/api/functions/index.html", "r+") as f:
         for line in f.readlines():
             if "add_columns" in line or "update_where" in line:
                 rendered_correctly = True
