--- tests/test_simple.py
@@ -1,2 +1,0 @@
-def test_simple():
-    assert 1 + 1 == 2
